refactor the service #2
