+++ src/main/java/de/hpi/shoprulesgenerator/service/ShopRulesGeneratorService.java
@@ -3,18 +3,11 @@ package de.hpi.shoprulesgenerator.service;
 import de.hpi.shoprulesgenerator.exception.ShopRulesDoNotExistException;
 import de.hpi.shoprulesgenerator.persistence.ShopRules;
 import de.hpi.shoprulesgenerator.persistence.repository.IShopRulesRepository;
-import de.hpi.shoprulesgenerator.properties.ShopRulesGeneratorConfig;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
 import org.springframework.stereotype.Service;
-
-import java.io.IOException;
-import java.util.Iterator;
-import java.util.List;
 
 @Slf4j
 @Service
@@ -22,13 +15,11 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ShopRulesGeneratorService implements IShopRulesGeneratorService {
 
+    private final HTMLPageFetcher htmlPageFetcher;
+
     private final IdealoBridge idealoBridge;
 
     private final IShopRulesRepository shopRulesRepository;
-
-    private final ShopRulesGeneratorConfig config;
-
-    private final URLCleaner urlCleaner;
 
     @Override
     public ShopRules getRules(long shopID) throws ShopRulesDoNotExistException {
@@ -42,39 +33,8 @@ public class ShopRulesGeneratorService implements IShopRulesGeneratorService {
 
     //actions
     private void generateRule(long shopID) {
-        List<IdealoOffer> idealoOffers = getIdealoBridge().getSampleOffers(shopID);
-        fetchHtmlPages(idealoOffers, shopID);
+        IdealoOffers idealoOffers = getIdealoBridge().getSampleOffers(shopID);
+        getHtmlPageFetcher().fetchHTMLPages(idealoOffers, shopID);
     }
 
-    private void fetchHtmlPages(List<IdealoOffer> idealoOffers, long shopID) {
-        for (Iterator<IdealoOffer> iterator = idealoOffers.iterator(); iterator.hasNext();) {
-            IdealoOffer offer = iterator.next();
-            fetchHtmlPage(offer, shopID);
-            if (iterator.hasNext()) {
-                try {
-                    Thread.sleep(getConfig().getFetchDelay());
-                } catch (InterruptedException e) {
-                    log.warn("Error while waiting between two fetches.", e);
-                    Thread.currentThread().interrupt();
-                }
-            }
-        }
-    }
-
-    private void fetchHtmlPage(IdealoOffer offer, long shopID) {
-        String cleanUrl = cleanUrl(offer.getOfferAttribute(OfferAttribute.URL), shopID);
-        try {
-            Document fetchedPage = Jsoup.connect(cleanUrl).userAgent(getConfig().getUserAgent()).get();
-            offer.setFetchedPage(fetchedPage);
-        } catch (IOException e) {
-            log.error("Could not fetch page for: " + cleanUrl, e);
-        }
-    }
-
-    private String cleanUrl(List<String> urls, long shopID) {
-        if (urls.isEmpty()) {
-            return null;
-        }
-        return getUrlCleaner().cleanURL(urls.get(0), shopID);
-    }
 }
